requirements.txt
flask==2.0.1
PyPDF2==1.26.0
python-docx==0.8.11
werkzeug==2.0.3